+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes
 * [Developer]: Updated FastQC to 0.11.7.
 * [UI]: Fixed bug in sample edit page that didn't allow users to clear a field in the sample.
 * [Workflow]: Updated the AssemblyAnnotation pipeline to v0.5 and the AssemblyAnnotationCollection pipeline to v0.4. Both pipelines now use Shovill for assembly and QUAST for assembly quality assessment in addition to Prokka for annotation.
+* [API]: Fixed REST endpoint mapping for current user and user projects.
 * [UI]: New project line list page with inline editing.
 
 0.21.0 to 0.22.0